+++ src/sentry/models/__init__.py
@@ -12,7 +12,7 @@ from .auditlogentry import *  # NOQA
 from .authenticator import *  # NOQA
 from .authidentity import *  # NOQA
 from .authprovider import *  # NOQA
-from .avatar import *  # NOQA
+from .avatars import *  # NOQA
 from .broadcast import *  # NOQA
 from .commit import *  # NOQA
 from .commitauthor import *  # NOQA
