+++ src/sentry/models/avatars/base.py
