+++ src/sentry/models/avatars/doc_integration_avatar.py
