+++ src/sentry/models/avatars/organization_avatar.py
