+++ src/sentry/models/avatars/project_avatar.py
