+++ src/sentry/models/avatars/sentry_app_avatar.py
