+++ src/sentry/models/avatars/team_avatar.py
