+++ src/sentry/models/avatars/user_avatar.py
